+++ realgui/3rd/Lite3D/widget/l3_sw_raster.h
@@ -105,7 +105,7 @@ typedef struct
 {
     uint8_t *writebuf;
     int write_off;
-    uint32_t image_base;
+    uintptr_t image_base;
     uint32_t image_off;
     char input_type;
     uint8_t dc_bytes_per_pixel;
@@ -120,7 +120,7 @@ bool l3_draw_img_target_area(l3_draw_rect_img_t *image, l3_canvas_t *dc, l3_rect
                              int32_t *x_start, int32_t *x_end, int32_t *y_start, int32_t *y_end);
 void gui_get_source_color(uint8_t *source_red, uint8_t *source_green, uint8_t *source_blue,
                           uint8_t *source_alpha,
-                          uint32_t image_base, uint32_t image_off, char input_type);
+                          uintptr_t image_base, uint32_t image_off, char input_type);
 
 #ifdef __cplusplus
 }
