+++ realgui/engine/sw/rle/acc_sw_rle.c
@@ -95,8 +95,8 @@ void uncompressed_rle_rect(imdc_file_t *file, int16_t x, int16_t y, int16_t w, i
 void uncompressed_rle_rgb565(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 {
     //imdc_file_header_t *header = (imdc_file_header_t *)file;
-    uint32_t start = (uint32_t)(uintptr_t)file + file->compressed_addr[line];
-    uint32_t end = (uint32_t)(uintptr_t)file + file->compressed_addr[line + 1];
+    uint32_t start = (uintptr_t)file + file->compressed_addr[line];
+    uint32_t end = (uintptr_t)file + file->compressed_addr[line + 1];
     uint16_t *linebuf = (uint16_t *)buf;
 
     // gui_log("file->compressed_addr[%d] %d\n", line, file->compressed_addr[line]);
@@ -113,8 +113,8 @@ void uncompressed_rle_rgb565(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 void uncompressed_rle_argb8565(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 {
     //imdc_file_header_t *header = (imdc_file_header_t *)file;
-    uint32_t start = (uint32_t)(uintptr_t)file + file->compressed_addr[line];
-    uint32_t end = (uint32_t)(uintptr_t)file + file->compressed_addr[line + 1];
+    uint32_t start = (uintptr_t)file + file->compressed_addr[line];
+    uint32_t end = (uintptr_t)file + file->compressed_addr[line + 1];
     uint8_t *linebuf = (uint8_t *)buf;
 
     for (uint32_t addr = start; addr < end;)
@@ -133,8 +133,8 @@ void uncompressed_rle_argb8565(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 void uncompressed_rle_rgb888(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 {
     //imdc_file_header_t *header = (imdc_file_header_t *)file;
-    uint32_t start = (uint32_t)(uintptr_t)file + file->compressed_addr[line];
-    uint32_t end = (uint32_t)(uintptr_t)file + file->compressed_addr[line + 1];
+    uint32_t start = (uintptr_t)file + file->compressed_addr[line];
+    uint32_t end = (uintptr_t)file + file->compressed_addr[line + 1];
     uint8_t *linebuf = (uint8_t *)buf;
 
     for (uint32_t addr = start; addr < end;)
@@ -153,8 +153,8 @@ void uncompressed_rle_rgb888(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 void uncompressed_rle_argb8888(imdc_file_t *file, uint32_t line,  uint8_t *buf)
 {
     //imdc_file_header_t *header = (imdc_file_header_t *)file;
-    uint32_t start = (uint32_t)(uintptr_t)file + file->compressed_addr[line];
-    uint32_t end = (uint32_t)(uintptr_t)file + file->compressed_addr[line + 1];
+    uint32_t start = (uintptr_t)file + file->compressed_addr[line];
+    uint32_t end = (uintptr_t)file + file->compressed_addr[line + 1];
     uint32_t *linebuf = (uint32_t *)buf;
 
     for (uint32_t addr = start; addr < end;)
@@ -179,7 +179,7 @@ void sw_acc_rle_uncompress(draw_img_t *image, void *buf)
 
     memcpy(buf, head, sizeof(struct gui_rgb_data_head));
     head = buf;
-    line_buf = (uint8_t *)(sizeof(struct gui_rgb_data_head) + (uint32_t)(uintptr_t)(buf));
+    line_buf = (uint8_t *)(sizeof(struct gui_rgb_data_head) + (uintptr_t)(buf));
     if (head->type == RGB565)//rle_rgb565
     {
         uint8_t source_bytes_per_pixel = 2;
